rfs: Proper type for size definitions

+++ include/samsung-ipc.h
@@ -113,8 +113,8 @@ char *ipc_client_nv_data_md5_path(struct ipc_client *client);
 char *ipc_client_nv_data_backup_path(struct ipc_client *client);
 char *ipc_client_nv_data_backup_md5_path(struct ipc_client *client);
 char *ipc_client_nv_data_secret(struct ipc_client *client);
-int ipc_client_nv_data_size(struct ipc_client *client);
-int ipc_client_nv_data_chunk_size(struct ipc_client *client);
+size_t ipc_client_nv_data_size(struct ipc_client *client);
+size_t ipc_client_nv_data_chunk_size(struct ipc_client *client);
 
 const char *ipc_response_type_to_str(int type);
 const char *ipc_request_type_to_str(int type);
